% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\name{geosse_tree}
\alias{geosse_tree}
\title{Geosse tree}
\format{
yaml
}
\description{
A description of the data stored in the YAML file.
}
\details{
This YAML file contains ...
}
\examples{
file_path <- system.file("extdata", "geosse_tb_tree.rds", package = "divextra")
phy <- readRDS(file_path)
}
