% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sse.R
\name{get_aic}
\alias{get_aic}
\title{Akaike Information Criterion (AIC)}
\usage{
get_aic(vec, npar)
}
\arguments{
\item{vec}{a vector of likelihood values}

\item{npar}{number of parameters}
}
\value{
AIC scores
}
\description{
The Akaike Information Criterion (AIC) is a measure used to compare the relative quality
of statistical models for a given dataset. It is calculated as:

\deqn{AIC = 2k - 2 \ln(L)}
}
\examples{

get_aic(2768.948, 4)
}
