% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sse.R
\name{init.pars.classe_td}
\alias{init.pars.classe_td}
\title{Creates a starting point for episodic ClaSSE with parameter constrains}
\usage{
init.pars.classe_td(lik.const.td, phy, k, n.epoch)
}
\arguments{
\item{lik.const.td}{constrained likelihood function}

\item{phy}{tree}

\item{k}{number of states}

\item{n.epoch}{number of epochs}
}
\value{
A named vector containing starting parameter values for each time epoch.
}
\description{
Creates a starting point for episodic ClaSSE with parameter constrains
}
\examples{

file_path <- system.file("extdata", "geosse_tb_tree.rds", package = "divextra")
phy <- readRDS(file_path)
file_path <- system.file("extdata", "geosse3_td_test.yml", package = "divextra")
par.categories.td <- read_yaml_pars_td(file_path)
formula.td <- make_constraints_sse_td(par.categories.td)
lik.td <-make.classe.td(phy, phy$tip.state, k=3, n.epoch=2, control=list(backend = "gslode"))
lik.const.td <- diversitree:::constrain(lik.td, formulae = formula.td)
starting.point <- init.pars.classe_td(lik.const.td, phy, k=3, n.epoch=2)
}
