% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe_td.R
\name{make.classe.td}
\alias{make.classe.td}
\title{Episodic ClaSSE model where different time epochs may have different parameters}
\usage{
make.classe.td(
  tree,
  states,
  k,
  n.epoch,
  sampling.f = NULL,
  strict = TRUE,
  control = list()
)
}
\arguments{
\item{tree}{An ultrametric bifurcating phylogenetic tree, in ape “phylo” format.}

\item{states}{A vector of character states, from 1 to Inf. This vector must have names that correspond to the tip labels in the phylogenetic tree (tree$tip.label).}

\item{k}{The number of states. The maximum now is 31, but that can easily be increased if necessary.}

\item{n.epoch}{Number of epochs. 1 corresponds to plain ClaSSE, so this will generally be an integer at least 2.}

\item{sampling.f}{sampling fraction}

\item{strict}{The states vector is always checked to make sure that the values are integers on 1:k.}

\item{control}{List of control parameters for the ODE solver. See details in diversitree:::make.bisse.}
}
\value{
a function
}
\description{
Prepare to run episodic ClaSSE on a phylogenetic tree and character states.
This function creates a likelihood function that can be used in ML or Bayesian inference.
}
\details{
This functions is based on diversitree::make.musse.td() and diversitree::make.classe().
}
\examples{

phy <- ape::read.tree(text = '(sp1:1.014858647,(((sp4:0.2614280769,sp5:0.2614280769)nd4:0.007869378189,sp3:0.2692974551)nd3:0.01878075832,sp2:0.2880782134)nd2:0.7267804339)nd1;')
states <- species_counts <- c(sp1 = 2, sp2 = 2, sp3 = 1, sp4 = 0, sp5 = 1)

# classe model
lik.c <- diversitree::make.classe(phy, states+1, 3)
pars.c <- diversitree::starting.point.classe(phy, 3)
lik.c(pars.c)   # -8.367568

# classe.td where two time bins have the same pars (=classe)
lik.ctd <- make.classe.td(phy, states + 1, k=3, n.epoch=2, control=list(backend = "gslode"))
par.ctd <- c(0.2, pars.c, pars.c)
names(par.ctd) <- diversitree::argnames(lik.ctd)
par.ctd
lik.ctd(par.ctd) # -8.367568

# classe.td where two time bins have different pars
lik.ctd2 <- make.classe.td(phy, states + 1, k=3, n.epoch=2, control=list(backend = "gslode"))
par.ctd2 <- c(0.2, pars.c, pars.c)
names(par.ctd2) <- diversitree::argnames(lik.ctd)
par.ctd2['lambda222.2'] <- 0.4
par.ctd2['lambda222.1'] <- 0.1
lik.ctd(par.ctd2) # -8.268788

}
