% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-fun.R
\name{make_initial_conditions_mine}
\alias{make_initial_conditions_mine}
\title{Title make intis}
\usage{
make_initial_conditions_mine(k)
}
\arguments{
\item{k}{param}
}
\value{
function
}
\description{
Title make intis
}
\examples{

make_initial_conditions_mine(3)
}
