% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-classe-td.R
\name{pars.geosse2classe}
\alias{pars.geosse2classe}
\title{Converts ClaSSE parameters to GeoSSE}
\usage{
pars.geosse2classe(pars.ge)
}
\arguments{
\item{pars.ge}{}
}
\value{
vector
}
\description{
It is based on diversitree:::pars.ge.to.cl(). Only works for three states
and time-homogeneous models. The state order is important: A, B, AB.
}
\examples{
## "sA"  "sB" "sAB" "xA"  "xB"  "dA"  "dB"
pars.ge <- matrix(
c(0.1, 0.1, 0.1,  0.001, 0.001,  0.1, 0.1,
  0.3, 0.3, 0.3,  0.001, 0.001,  0.1, 0.1
),
2, 7, byrow = TRUE)
colnames(pars.ge) <- diversitree:::default.argnames.geosse()
pars.cl <- t(apply(pars.ge, 1, pars.geosse2classe))
print(pars.cl)
}
