% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sse.R
\name{pars2groups}
\alias{pars2groups}
\title{Group parameters based on the same parameter estimates}
\usage{
pars2groups(vec)
}
\arguments{
\item{vec}{A named vector of parameter estimates}
}
\value{
a list
}
\description{
Group parameters based on the same parameter estimates
}
