% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sse.R
\name{pars_to_arrays}
\alias{pars_to_arrays}
\title{Display ClaSSE parameters as arrays}
\usage{
pars_to_arrays(pars, Nstates, nm = NULL)
}
\arguments{
\item{pars}{ClaSSE parameters}

\item{Nstates}{number of character states}

\item{nm}{names for character states}
}
\value{
a list of arrays
}
\description{
Takes a named vector of ClaSSE parameters and displays them as arrays
}
\examples{

reg <- c("A", "B", "A.B")
pars <- diversitree:::default.argnames.classe(3)
args <- pars_to_arrays(pars, 3, reg)
print(args)
}
