% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-yaml.R
\name{pars_yaml_to_arrays}
\alias{pars_yaml_to_arrays}
\title{Display symbolic parameters of time-homogeneous ClaSSE as arrays}
\usage{
pars_yaml_to_arrays(par.categories)
}
\arguments{
\item{par.categories}{named list of parameters and configurations (e.g., from yaml file)}
}
\value{
a list of arrays
}
\description{
Display symbolic parameters of time-homogeneous ClaSSE as arrays
}
\examples{

file_path <- system.file("extdata", "geosse3_ti_maxpar.yml", package = "divextra")
yaml_pars <- read_yaml_pars(file_path)
pars_yaml_to_arrays(yaml_pars)
}
