% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-yaml.R
\name{pars_yaml_to_arrays_td}
\alias{pars_yaml_to_arrays_td}
\title{Display symbolic parameters of episodic ClaSSE as arrays}
\usage{
pars_yaml_to_arrays_td(par.categories.td)
}
\arguments{
\item{par.categories.td}{named list of parameters and configurations (e.g., from yaml file)}
}
\value{
a list of arrays
}
\description{
Display symbolic parameters of episodic ClaSSE as arrays
}
\examples{

file_path <- system.file("extdata", "geosse3_td_test.yml", package = "divextra")
par.categories.td <- read_yaml_pars_td(file_path)
par.symbolic.td <- pars_yaml_to_arrays_td(par.categories.td)
print(par.categories.td)
print(par.symbolic.td)
}
