% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{phy5}
\alias{phy5}
\title{Toy Tree Data}
\format{
phylo object
}
\source{
Generated using diversitree simulation functions
}
\usage{
phy5
}
\description{
A phylogenetic tree with five tips
}
\examples{
data(phy5)
print(phy5)

}
\keyword{datasets}
