% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-yaml.R
\name{read_yaml_pars}
\alias{read_yaml_pars}
\title{Read parameters of time-homogeneous GeoSSE model from yaml file}
\usage{
read_yaml_pars(yaml_file)
}
\arguments{
\item{yaml_file}{yaml file}
}
\value{
named list of parameters and configurations
}
\description{
Read parameters of time-homogeneous GeoSSE model from yaml file
}
\examples{

file_path <- system.file("extdata", "geosse3_ti_maxpar.yml", package = "divextra")
par.categories <- read_yaml_pars(file_path)
print(par.categories)
}
