% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-yaml.R
\name{read_yaml_pars_td}
\alias{read_yaml_pars_td}
\title{Read parameters of episodic GeoSSE model from yaml file}
\usage{
read_yaml_pars_td(yaml_file)
}
\arguments{
\item{yaml_file}{yaml file}
}
\value{
named list of parameters and configurations
}
\description{
Read parameters of episodic GeoSSE model from yaml file
}
\examples{

file_path <- system.file("extdata", "geosse3_td_maxpar.yml", package = "divextra")
par.categories.td <- read_yaml_pars_td(file_path)
print(par.categories.td)
}
