% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-yaml.R
\name{regroup_parameters}
\alias{regroup_parameters}
\title{Regroup Parameters According to New Groups}
\usage{
regroup_parameters(yaml_pars, new_groups)
}
\arguments{
\item{yaml_pars}{List of original parameter groupings}

\item{new_groups}{List of vectors containing group names to combine}
}
\value{
List of new parameter groupings
}
\description{
Takes original parameter grouping and creates new parameter groups based on provided grouping scheme.
Checks that new groups don't have overlapping parameters.
}
\examples{
\dontrun{
# Original parameter groups from model
orig_pars <- list(pars =list(
  dAE = c("q0107.1", "q0107.2"),
  dAM = c("q0108.1", "q0108.2"),
  scA = c("lambda070102.1", "lambda070102.2"),
  scE = c("lambda120203.1", "lambda120203.2")
)
)

# New grouping scheme
new_groups <- list(
  c("dAE", "dAM"),
  c("scA", "scE")
)

regroup_parameters(orig_pars, new_groups)
}
}
