% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sse.R
\name{reoder_pars}
\alias{reoder_pars}
\title{Reorder character states in arrays}
\usage{
reoder_pars(pars.array, v)
}
\arguments{
\item{pars.array}{a list of arrays}

\item{v}{new order of character states}
}
\value{
a list of arrays
}
\description{
Reorder character states in arrays
}
\examples{

reg <- c("A", "B", "A.B")
pars <- diversitree:::default.argnames.classe(3)
args <- pars_to_arrays(pars, 3, reg)
print(args)
new.order <- reoder_pars(args, c(3,1,2))
print(new.order)
}
