% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-fun.R
\name{report_p}
\alias{report_p}
\title{Report p-value}
\usage{
report_p(p, digits = 3)
}
\arguments{
\item{p}{The p-value}

\item{digits}{The number of digits to round to (default = 3)}
}
\value{
A string with the format "p = .040" or "p < .001"
}
\description{
Reports a p-value in APA style.
}
\examples{

report_p(0.02018) # returns "p = .020"
report_p(0.00028) # returns "p < .001"

}
