% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-classe-td.R
\name{table2tree}
\alias{table2tree}
\title{Converts episodic ClaSSE.td simulations to a tree}
\usage{
table2tree(info)
}
\arguments{
\item{info}{a table from make.tree.classe.td()}
}
\value{
tree
}
\description{
Converts an output (a table) from make.tree.classe.td() to a phylogenetic tree
that contains extant species only. It is based on diversitree:::me.to.ape.bisse().
}
\examples{

## "sA"  "sB" "sAB" "xA"  "xB"  "dA"  "dB"
pars.ge <- matrix(
  c(0.1, 0.1, 0.1,  0, 0,  0.1, 0.1,
    0.3, 0.3, 0.3,  0, 0,  0.1, 0.1
  ),
  2, 7, byrow = TRUE)
colnames(pars.ge) <- diversitree:::default.argnames.geosse()
pars.cl <- t(apply(pars.ge, 1, pars.geosse2classe))
print(pars.cl)

set.seed(123)
tb <- make.tree.classe.td(pars.cl, k=3, max.t1=10, max.t2=15, x0=1, single.lineage=TRUE)
print(tb)
phy <- table2tree(tb)
plot(phy)

}
