% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\name{yaml_files}
\alias{yaml_files}
\title{Example YAML Data: My Data NEW}
\format{
yaml
}
\description{
A description of the data stored in the YAML file.
}
\details{
This YAML file contains ...
}
\examples{
file_path <- system.file("extdata", "geosse3_td_test.yml", package = "divextra")
data <- yaml::read_yaml(file_path)
}
