% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-classe-td.R
\name{make.tree.classe.td}
\alias{make.tree.classe.td}
\title{Simulate episodic ClaSSE tree and character states}
\usage{
make.tree.classe.td(
  pars.tb,
  k,
  max.taxa = Inf,
  max.t1 = Inf,
  max.t2 = Inf,
  x0,
  single.lineage = TRUE
)
}
\arguments{
\item{pars.tb}{parameters, a matrix with two rows with each row correponding to one epoch}

\item{k}{number of states}

\item{max.taxa}{Maximum number of taxa to include in the tree. If Inf, then the tree will be evolved until max.t time has passed.}

\item{max.t1}{Maximum tree length to evolve during epoch 1.}

\item{max.t2}{Maximum tree length to evolve in total (i.e., epoch 1 + epoch 2)}

\item{x0}{Initial character state at the root (1, .., k)}

\item{single.lineage}{Start simulation with a single lineage?}
}
\value{
data frame
}
\description{
It is based on diversitree:::make.tree.classe() but simulates an episodic (time-dependent) ClaSSE tree
where the tree is split into two time epochs, each with its own parameters.
}
\details{
The time epochs start from the root, so the simulation begins with epoch 1 and ends in epoch 2.
The maximum number of epochs is two.' Note, in ML calculation with make.classe.td(),
the epoch order is reversed -- epoch 1 starts at tips.

The simulations are tested only with max.t1 and max.t2 arguments. Use max.taxa at your own risk.
The argument single.lineage should be always set to TRUE.

The returned data frame contains three attributes: attr(info, "info.epoch1") attr(info, "t.epoch1") attr(info, "t.total").
}
\examples{

## "sA"  "sB" "sAB" "xA"  "xB"  "dA"  "dB"
pars.ge <- matrix(
  c(0.1, 0.1, 0.1,  0, 0,  0.1, 0.1,
    0.3, 0.3, 0.3,  0, 0,  0.1, 0.1
  ),
  2, 7, byrow = TRUE)
colnames(pars.ge) <- diversitree:::default.argnames.geosse()
pars.cl <- t(apply(pars.ge, 1, pars.geosse2classe))
print(pars.cl)

set.seed(123)
tb <- make.tree.classe.td(pars.cl, k=3, max.t1=10, max.t2=15, x0=1, single.lineage=TRUE)
print(tb)
phy <- table2tree(tb)
plot(phy)

# the tree from epoch 1 only
tb.e1 <-  attr(tb, "info.epoch1")
phy.e1 <- diversitree:::me.to.ape.bisse(tb.e1[-1,], tb.e1$state[1])
phy.e1 <- diversitree::prune(phy.e1) # prune extinct
plot(phy.e1)

# the final tree (after epoch 2)
phy$t.epoch1 # time of epoch 1 (=max.t1)
phy$t.total # total time (=max.t2)
phy$epoch1.extant.tree.depth # tree depth from epoch 1
phy$epoch12.tree.depth # total tree depth
phy$epoch1.ntip # N of tips in epoch 1 only
ape::Ntip(phy) # total N of tips

# The model changes regimes at (if counting from the tips for the infference):
phy$t.regime.change # := phy$epoch12.tree.depth - phy$epoch1.extant.tree.depth
}
